research/claude-task-runner/backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pyyaml>=6.0
aiofiles>=23.2.1
python-multipart>=0.0.6
